+++ src/simpleddp.js
@@ -380,20 +380,21 @@ class simpleDDP {
 
 	/**
 	 * Tries to subscribe to a specific publication on server.
+	 * Restarts the subscription if the same subscription exists.
 	 * @public
 	 * @param {string} pubname - Name of the publication on server.
 	 * @param {Array} [arguments] - Array of parameters to pass to the remote method. Pass an empty array or don't pass anything if you do not wish to pass any parameters.
 	 * @return {ddpSubscription} - Subscription.
 	 */
 	sub(pubname,args) {
-		let hasSuchSub = this.subs.find((sub)=>{
+		let hasSuchSub = this.subs.find((sub) => {
 			return sub.pubname == pubname && isEqual(sub.args,Array.isArray(args)?args:[]);
 		});
 		if (!hasSuchSub) {
 			let i = this.subs.push(new ddpSubscription(pubname,Array.isArray(args)?args:[],this));
 			return this.subs[i-1];
 		} else {
-			// perhaps the sub can be restarted here just in case
+			hasSuchSub.restart();
 			return hasSuchSub;
 		}
 	}